requirements.txt
annotated-types==0.6.0
anyio==4.3.0
bcc==0.18.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
chardet==4.0.0
chrome-gnome-shell==0.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
exceptiongroup==1.2.0
gorilla-cli==0.0.10
h11==0.14.0
halo==0.0.31
hidpidaemon==18.4.6
httpcore==1.0.4
httplib2==0.20.2
httpx==0.27.0
idna==3.3
importlib-metadata==4.6.4
jeepney==0.7.1
kernelstub==3.1.4
keyring==23.5.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
log-symbols==0.0.14
louis==3.20.0
macaroonbakery==1.3.1
more-itertools==8.10.0
netaddr==0.8.0
netifaces==0.11.0
oauthlib==3.2.0
openai==1.14.2
pop-transition==1.1.2
prompt-toolkit==3.0.43
protobuf==3.12.4
pycairo==1.20.1
pycups==2.0.1
pydantic==2.6.4
pydantic_core==2.16.3
pydbus==0.6.0
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu3
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-gnupg==0.4.8
python-xlib==0.29
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
repolib==2.2.1
repoman==1.4.0
requests==2.25.1
SecretStorage==3.3.1
sessioninstaller==0.0.0
six==1.16.0
sniffio==1.3.1
spinners==0.0.24
systemd-python==234
termcolor==2.4.0
tqdm==4.66.2
typing_extensions==4.10.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
urllib3==1.26.5
vboxapi==1.0
wadllib==1.3.6
wcwidth==0.2.13
xdg==5
xkit==0.0.0
zipp==1.0.0
